Implemented delete operation for Book entity.

+++ src/BookstoreEJB/src/main/java/com/github/cvetan/bookstore/sb/book/BookSB.java
@@ -1,6 +1,7 @@
 package com.github.cvetan.bookstore.sb.book;
 
 import com.github.cvetan.bookstore.exceptions.book.BookIsbnUsedException;
+import com.github.cvetan.bookstore.exceptions.book.BookOrderFKException;
 import com.github.cvetan.bookstore.model.Book;
 import com.github.cvetan.bookstore.sb.BookstoreSB;
 import com.github.cvetan.bookstore.util.SlugGenerator;
@@ -48,10 +49,12 @@ public class BookSB extends BookstoreSB implements BookSBLocal {
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(int id) throws BookOrderFKException {
         Book book = (Book) entityManager.find(Book.class, id);
         
-        // FK checks TODO
+        if (book.getOrderItemList().size() > 0) {
+            throw new BookOrderFKException("bookBoughtError");
+        }
         
         entityManager.remove(book);
